Fix failed queue hrefs when using later version of Resque
Strap yourself in.

This line needs to be changed in order to use this gem in a Rails 5 app
(with Resque 1.22.0) and have working views.

Before this commit: this gem is known to work without issues with Resque
1.13.0 in a Rails 3.2 app. However:
- That version of resque has a dependency on json 1.4.6
- Ruby 2.2 is only compatible with > json 1.8
- Rails 5 is only compatible with >= ruby 2.2.2

Resque 1.22.0 is the highest version that removes the json 1.4.6
dependency whilst keeping fully working views for multiple failure
queues.

Despite newer versions of the Resque gem actually having the code from
this gem merged in - it doesn't appear that enough people use the
multiple failure queues behaviour because they haven't kept it working.
The newer resque-web gem also doesn't have working views for multiple
failure queues.